+++ Aether/core/src/utils/text_format.rs
@@ -98,21 +98,21 @@ pub const MCP_RESULT_TRUNCATE_LENGTH: usize = 2000;
 /// Get the default text truncate length from policy or use constant
 pub fn get_default_truncate_length(policy: Option<&TextFormatPolicy>) -> usize {
     policy
-        .map(|p| p.default_truncate_length)
+        .map(|p| p.default_truncate_length as usize)
         .unwrap_or(DEFAULT_TEXT_TRUNCATE_LENGTH)
 }
 
 /// Get the search snippet truncate length from policy or use constant
 pub fn get_search_snippet_length(policy: Option<&TextFormatPolicy>) -> usize {
     policy
-        .map(|p| p.search_snippet_length)
+        .map(|p| p.search_snippet_length as usize)
         .unwrap_or(SEARCH_SNIPPET_TRUNCATE_LENGTH)
 }
 
 /// Get the MCP result truncate length from policy or use constant
 pub fn get_mcp_result_length(policy: Option<&TextFormatPolicy>) -> usize {
     policy
-        .map(|p| p.mcp_result_length)
+        .map(|p| p.mcp_result_length as usize)
         .unwrap_or(MCP_RESULT_TRUNCATE_LENGTH)
 }
